Update README.md
Facial recognition is a smart technology that helps computers identify or verify a person’s identity using their face. It works by capturing an image, detecting a face, analyzing unique features like eyes, nose, and mouth, and then comparing it with stored data to find a match.

This technology is used in many areas, such as unlocking smartphones, improving security in airports, tracking attendance in schools and offices, and even in smart cameras for personalized experiences. It is fast, convenient, and widely used to enhance security and automation in daily life.